+++ src/request.rs
@@ -78,7 +78,7 @@ impl<Body> SentRequest<Body> {
 	///
 	/// This function returns `Ok(None)` if the final response is received instead of an update message.
 	/// If that happens, the response message can be read using [`Self::response`].
-	pub async fn recv_update(&mut self) -> Result<Option<Message<Body>>, error::NextMessageError> {
+	pub async fn recv_update(&mut self) -> Result<Option<Message<Body>>, error::RecvMessageError> {
 		let message = self.recv_message().await?;
 		if message.header.message_type.is_responder_update() {
 			Ok(Some(message))
@@ -93,7 +93,7 @@ impl<Body> SentRequest<Body> {
 	/// This function returns an [`InvalidMessageType`][error::InvalidMessageType] if the received message is an update message.
 	/// If that happens, the update message can be read using [`Self::next_update`].
 	/// To ensure that there are no update messages left, keep calling [`Self::next_update`] untill it returns `Ok(None)`.
-	pub async fn recv_response(&mut self) -> Result<Message<Body>, error::NextMessageError> {
+	pub async fn recv_response(&mut self) -> Result<Message<Body>, error::RecvMessageError> {
 		let message = self.recv_message().await?;
 		let kind = message.header.message_type;
 		if kind.is_response() {
@@ -110,7 +110,7 @@ impl<Body> SentRequest<Body> {
 	/// Receive the next message of the request from the remote peer.
 	///
 	/// This could be an update message or a response message.
-	async fn recv_message(&mut self) -> Result<Message<Body>, error::NextMessageError> {
+	async fn recv_message(&mut self) -> Result<Message<Body>, error::RecvMessageError> {
 		if let Some(message) = self.peek_buffer.take() {
 			Ok(message)
 		} else {